Start webpack-dev-server via `yarn start`
The webpack-dev-server enables automatic recompilation of JS and SCSS
files and refreshing of the browser. This means that when you edit e.g.
a `.jsx` file, you don't have to hit reload in the browser; the changes
will just magically appear.

The README previously suggested to run `bin/webpack-dev-server` in a
separate terminal session. This commit simplifies this by introducing a
`Procfile.dev` that starts both the Rails server and the webpack dev
server. The `yarn start` shortcut now does this:

    heroku local -f Procfile.dev

This makes it much quicker for a developer to get the app up and running
in a single terminal session with the live reload behavior for webpack.